Major refactoring of the code to check for categories as well as galleries.
Since a single file may be referenced by a gallery and a category (or multiple categories) we also need to implement a scoring system, to pick the best taxonomy for a file.

Obtaining the categories requires adjusting the current call to the Wikimedia API so that it returns categories in batches of 50 pages. Hence this commit also changes the way the SchemaDocument is saved, so that it is saved incrementally.