Save Flickr rotation info in the "additionalInformation" field of the dataObject, then pass it in to grab_file so that the image can be correctly rotated when harvesting.
This will help with a number of flickr images that have been rotated by hand in Flickr (many BHL images).

The flickr "rotation" parameter includes any rotations specified in the EXIF metadata, so if we have a Flickr rotation value, we should override the EXIF rotation specification.

I'm unsure of the correct way to access the additionalInformation field from within the DataObjects class - a comment is inserted as appropriate. This will probably need changing.