PHP tests now pass without eol-private
Currently it essentially turns off the Flickr tests since they rely on
resolving Flickr URLs.  Ideally this would be handled with web mocks,
but the current PHP testing infrastructure we are using, SimpleTest,
doesn't support this.  We should either change to PHPUnit or convert
all the harvesting code over to Ruby.