Change calling convention: all crops done at once
To save the crop location, we need the crop code to return a single set
of values, even if there are multiple square thumbnail sizes. This
change means that all the cropping is done by a single call to the
(renamed) function create_crops(), which returns the percentage crop
positions.